adjust setting dlg. fix soui3 bug. update ver to 3.0.5.45

+++ isserver3/IsSvrProxy.cpp
@@ -213,7 +213,7 @@ void CIsSvrProxy::_OnBuildIndexProg(int indexMode, PROGTYPE uType, unsigned int 
 	if (uType == PT_MAX) nMax = dwData, nPrevProg = 0;
 	if (uType == PT_PROG)
 	{
-		if (dwData - nPrevProg < nMax / 200)
+		if ((int)dwData - nPrevProg < nMax / 200)
 		{
 			return;
 		}
